Remove BaseGame class
Everything inherits from the Game class anyway, and all "abstract"
methods are already implemented there.
