+++ server/abc/base_game.py
@@ -1,4 +1,3 @@
-from abc import ABCMeta, abstractmethod
 from enum import Enum, IntEnum, unique
 
 
@@ -14,38 +13,3 @@ class GameConnectionState(Enum):
 class InitMode(IntEnum):
     NORMAL_LOBBY = 0
     AUTO_LOBBY = 1
-
-
-class BaseGame:
-    __metaclass__ = ABCMeta
-
-    @abstractmethod
-    async def on_game_end(self):
-        pass  # pragma: no cover
-
-    async def rate_game(self):
-        pass  # pragma: no cover
-
-    @property
-    @abstractmethod
-    def init_mode(self):
-        """
-        The initialization mode to use for the Game
-        :rtype InitMode
-        :return:
-        """
-        pass  # pragma: no cover
-
-    @abstractmethod
-    def teams(self):
-        """
-        A dictionary of lists representing teams
-
-        It is of the form:
-        >>> {
-        >>>     1: [Player(1), Player(2)],
-        >>>     2: [Player(3), Player(4)]
-        >>> }
-        :return:
-        """
-        pass  # pragma: no cover
